Update BrandManageService class

+++ shop/services/manage/shop/BrandManageService.php
@@ -4,6 +4,8 @@ namespace shop\services\manage\shop;
 
 use shop\entities\Meta;
 use shop\entities\shop\Brand;
+use shop\entities\shop\product\Product;
+use shop\exceptions\NotFoundException;
 use shop\forms\manage\shop\BrandForm;
 use yii\helpers\Inflector;
 
@@ -15,9 +17,9 @@ class BrandManageService
             $form->name,
             $form->slug ?: Inflector::slug($form->name),
             new Meta(
-                $form->meta->title,
-                $form->meta->description,
-                $form->meta->keywords
+                $form->meta->title ?? null,
+                $form->meta->description ?? null,
+                $form->meta->keywords ?? null
             )
         );
 
@@ -28,15 +30,15 @@ class BrandManageService
 
     public function edit($id, BrandForm $form)
     {
-        $brand = Brand::findOne(['id' => $id]);
+        $brand = $this->getBrand($id);
 
         $brand->edit(
             $form->name,
             $form->slug,
             new Meta(
-                $form->meta->title,
-                $form->meta->description,
-                $form->meta->keywords
+                $form->meta->title ?? null,
+                $form->meta->description ?? null,
+                $form->meta->keywords ?? null
             )
         );
 
@@ -45,8 +47,21 @@ class BrandManageService
 
     public function remove($id)
     {
-        $brand = Brand::findOne(['id' => $id]);
+        $brand = $this->getBrand($id);
+
+        if (Product::find()->andWhere(['brand_id' => $id])->exists()) {
+            throw new \DomainException('Unable to delete brand with products.');
+        }
 
         $brand->delete();
+    }
+
+    private function getBrand($id) : Brand
+    {
+        if (!$brand = Brand::findOne(['id' => $id])) {
+            throw new NotFoundException('Brand not found.');
+        }
+
+        return $brand;
     }
 }
